venvが起動中かどうか判別するロジックを追加

+++ pip-remake.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
+# 仮想環境がアクティブかどうかを確認
+if [ -z "$VIRTUAL_ENV" ]; then
+    echo "Error: No active virtual environment found."
+    exit 1
+fi
+
 # packageをinstallし直す
 pip freeze > uninstall.txt
 pip uninstall -y -r uninstall.txt
